Refactor Auth/MemberServices
