+++ backend/src/main/java/com/hot6/pnureminder/service/AuthService.java
@@ -1,6 +1,10 @@
 package com.hot6.pnureminder.service;
 
-import com.hot6.pnureminder.dto.*;
+import com.hot6.pnureminder.dto.Member.LoginDto;
+import com.hot6.pnureminder.dto.Member.MemberRequestDto;
+import com.hot6.pnureminder.dto.Member.MemberResponseDto;
+import com.hot6.pnureminder.dto.Token.TokenReqDto;
+import com.hot6.pnureminder.dto.Token.TokenResDto;
 import com.hot6.pnureminder.entity.Member;
 import com.hot6.pnureminder.entity.RefreshToken;
 import com.hot6.pnureminder.entity.Role;
@@ -30,6 +34,7 @@ public class AuthService {
     private final RefreshTokenRepository refreshTokenRepository;
 
     private final SmtpEmailService smtpEmailService;
+    private final MemberService memberService;
 
     @Transactional
     public MemberResponseDto signup(MemberRequestDto memberRequestDto) {
@@ -49,7 +54,7 @@ public class AuthService {
     }
 
     @Transactional
-    public TokenDto login(LoginDto loginDto) {
+    public TokenResDto login(LoginDto loginDto) {
         // 1. Login ID/PW 를 기반으로 AuthenticationToken 생성
         UsernamePasswordAuthenticationToken authenticationToken = loginDto.toAuthentication();
 
@@ -58,24 +63,24 @@ public class AuthService {
         Authentication authentication = authenticationManagerBuilder.getObject().authenticate(authenticationToken);
 
         // 3. 인증 정보를 기반으로 JWT 토큰 생성
-        TokenDto tokenDto = tokenProvider.generateTokenDto(authentication);
+        TokenResDto tokenResDto = tokenProvider.generateTokenDto(authentication);
 
         // 4. RefreshToken 저장
         RefreshToken refreshToken = RefreshToken.builder()
             .key(authentication.getName())
-            .value(tokenDto.getRefreshToken())
+            .value(tokenResDto.getRefreshToken())
             .build();
 
         refreshTokenRepository.save(refreshToken);
 
         // 5. 토큰 발급
-        return tokenDto;
+        return tokenResDto;
     }
 
     @Transactional
-    public void logout(TokenRequestDto tokenRequestDto) {
+    public void logout(TokenReqDto tokenReqDto) {
         // 로그아웃하려는 사용자의 정보를 가져옴
-        Authentication authentication = tokenProvider.getAuthentication(tokenRequestDto.getAccessToken());
+        Authentication authentication = tokenProvider.getAuthentication(tokenReqDto.getAccessToken());
 
         // 저장소에서 해당 사용자의 refresh token 삭제
         refreshTokenRepository.deleteByKey(authentication.getName());
@@ -83,46 +88,50 @@ public class AuthService {
 
 
     @Transactional
-    public TokenDto reissue(TokenRequestDto tokenRequestDto) {
+    public TokenResDto reissue(TokenReqDto tokenReqDto) {
         // 1. Refresh Token 검증
-        if (!tokenProvider.validateToken(tokenRequestDto.getRefreshToken())) {
+        if (!tokenProvider.validateToken(tokenReqDto.getRefreshToken())) {
             throw new RuntimeException("Refresh Token 이 유효하지 않습니다.");
         }
 
         // 2. Access Token 에서 Member ID 가져오기
-        Authentication authentication = tokenProvider.getAuthentication(tokenRequestDto.getAccessToken());
+        Authentication authentication = tokenProvider.getAuthentication(tokenReqDto.getAccessToken());
 
         // 3. 저장소에서 Member ID 를 기반으로 Refresh Token 값 가져옴
         RefreshToken refreshToken = refreshTokenRepository.findByKey(authentication.getName())
             .orElseThrow(() -> new RuntimeException("로그아웃 된 사용자입니다."));
 
         // 4. Refresh Token 일치하는지 검사
-        if (!refreshToken.getValue().equals(tokenRequestDto.getRefreshToken())) {
+        if (!refreshToken.getValue().equals(tokenReqDto.getRefreshToken())) {
             throw new RuntimeException("토큰의 유저 정보가 일치하지 않습니다.");
         }
 
         // 5. 새로운 토큰 생성
-        TokenDto tokenDto = tokenProvider.generateTokenDto(authentication);
+        TokenResDto tokenResDto = tokenProvider.generateTokenDto(authentication);
 
         // 6. 저장소 정보 업데이트
-        RefreshToken newRefreshToken = refreshToken.updateValue(tokenDto.getRefreshToken());
+        RefreshToken newRefreshToken = refreshToken.updateValue(tokenResDto.getRefreshToken());
         refreshTokenRepository.save(newRefreshToken);
 
         // 토큰 발급
-        return tokenDto;
+        return tokenResDto;
     }
 
     @Transactional
-    public void issueTempPassword(String username) {
-        Member member = memberRepository.findByUsername(username)
-                .orElseThrow(() -> new RuntimeException("사용자를 찾을수 없습니다"));
+    public void issueTempPassword(String username, String nickname) {
+        Member member = memberService.findMemberByUsername(username);
+
+        if(!member.getNickname().equals(nickname)){
+            throw new RuntimeException("입력한 닉네임이 사용자의 닉네임과 일치하지 않습니다.");
+        }
 
         String tempPassword = generateTempPassword();
         member.setPassword(passwordEncoder.encode(tempPassword));
         memberRepository.save(member);
 
-        smtpEmailService.sendTempPassword(member.getUsername(),tempPassword);
+        smtpEmailService.sendTempPassword(member.getUsername(), tempPassword);
     }
+
 
     private String generateTempPassword() {
         return UUID.randomUUID().toString().substring(0, 8);
