+++ backend/src/main/java/com/hot6/pnureminder/service/CustomUserDetailsService.java
@@ -2,7 +2,7 @@ package com.hot6.pnureminder.service;
 
 import com.hot6.pnureminder.entity.Member;
 import com.hot6.pnureminder.repository.MemberRepository;
-import com.hot6.pnureminder.dto.MemberResponseDto;
+import com.hot6.pnureminder.dto.Member.MemberResponseDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
