+++ backend/src/main/java/com/hot6/pnureminder/service/MemberService.java
@@ -1,7 +1,6 @@
 package com.hot6.pnureminder.service;
 
-import com.hot6.pnureminder.dto.MemberFindIdReqDto;
-import com.hot6.pnureminder.dto.MemberResponseDto;
+import com.hot6.pnureminder.dto.Member.MemberResponseDto;
 import com.hot6.pnureminder.entity.Member;
 import com.hot6.pnureminder.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
